requirements.txt
openai
streamlit
pandas
yfinance
numpy
matplotlib
scikit-learn
python-dotenv
